Refactor joining code into lib/auth/join
auth.Register and auth.RegisterParams were moved into the new package
to prevent clients(tbot) from depending on lib/auth. For the moment
only iam joining has been moved into the new package. All other
join methods have been left behind in their lib/auth/join_foo.go
file and will be addressed in future PRs if need be.  This removes
lib/auth entirely from tbot and reduces its binary size by ~10-20MB.
